move dsn function to config
