+++ gorm/gorm.go
@@ -1,9 +1,6 @@
 package gorm
 
 import (
-	"fmt"
-	"os"
-
 	_ "github.com/lib/pq"
 	streampets "github.com/streampets/backend"
 	"gorm.io/driver/postgres"
@@ -13,18 +10,6 @@ import (
 type DB struct {
 	*gorm.DB
 	dsn string
-}
-
-// TODO: Not sure where to put this, maybe in config?
-func GET_DSN() string {
-	host := mustGetEnv("DB_HOST")
-	port := mustGetEnv("DB_PORT")
-	sslMode := mustGetEnv("DB_SSL_MODE")
-	dbName := mustGetEnv("DB_NAME")
-	user := mustGetEnv("DB_USER")
-	password := mustGetEnv("DB_PASSWORD")
-
-	return fmt.Sprintf("host=%s user=%s password=%s dbname=%s port=%s sslmode=%s", host, user, password, dbName, port, sslMode)
 }
 
 func NewDB(dsn string) *DB {
@@ -51,12 +36,4 @@ func (db *DB) Open() (err error) {
 	}
 
 	return nil
-}
-
-func mustGetEnv(name string) string {
-	value := os.Getenv(name)
-	if value == "" {
-		panic(fmt.Errorf("%s not set", name))
-	}
-	return value
 }
